requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.45.0
django-bootstrap-v5==1.0.5
django-heroku==0.3.1
google-api-core==2.2.2
google-api-python-client==2.29.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.2
importlib-metadata==3.10.1
oauthlib==3.1.1
protobuf==3.19.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.2.0
pyparsing==3.0.4
python-decouple==3.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.5.0
django-storages==1.12.3
boto3==1.20.6
botocore==1.23.6
